+++ vice/src/maincpu.c
@@ -387,7 +387,7 @@ inline static int interrupt_check_irq_delay(interrupt_cpu_status_t *cs,
 /* ------------------------------------------------------------------------- */
 
 #ifdef NEED_REG_PC
-unsigned int reg_pc;
+WORD reg_pc;
 #endif
 
 static BYTE **o_bank_base;
